back/requirements.txt
asgiref==3.7.2
attrs==23.1.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
Django==4.1.11
django-cookies-samesite==0.1.2
django-cors-headers==4.2.0
django-sslserver==0.22
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
firebase-admin==6.1.0
fonttools==4.39.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
kiwisolver==1.4.4
lxml==4.9.3
Markdown==3.4.4
matplotlib==3.7.1
msgpack==1.0.5
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
platformdirs==3.11.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==4.5.0
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
spyne==2.14.0
sqlparse==0.2.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
zeep==4.2.1
